spline #108 refactoring step 7: plugin auto discovery and dynamic loading. Get rid of a "secondary" logicalRelation handler
