+++ core/src/main/resources/spline.default.properties
@@ -32,6 +32,3 @@ spline.iwd_strategy.default.on_missing_metrics=IGNORE_LINEAGE
 
 # User extra metadata provider
 spline.user_extra_meta_provider.className=za.co.absa.spline.harvester.extra.NoopUserExtraMetaDataProvider
-
-# Proprietary relation support
-spline.read_relation_handler.className=za.co.absa.spline.harvester.builder.read.NoOpReadRelationHandler
