+++ core/src/main/scala/za/co/absa/spline/harvester/LineageHarvester.scala
@@ -32,7 +32,7 @@ import za.co.absa.spline.harvester.ExtraMetadataImplicits._
 import za.co.absa.spline.harvester.LineageHarvester._
 import za.co.absa.spline.harvester.ModelConstants.{AppMetaInfo, ExecutionEventExtra, ExecutionPlanExtra}
 import za.co.absa.spline.harvester.builder._
-import za.co.absa.spline.harvester.builder.read.{PluggableReadCommandExtractor, ReadRelationHandler}
+import za.co.absa.spline.harvester.builder.read.PluggableReadCommandExtractor
 import za.co.absa.spline.harvester.builder.write.PluggableWriteCommandExtractor
 import za.co.absa.spline.harvester.conf.SplineConfigurer.SplineMode
 import za.co.absa.spline.harvester.conf.SplineConfigurer.SplineMode.SplineMode
@@ -50,8 +50,7 @@ class LineageHarvester(
   hadoopConfiguration: Configuration,
   splineMode: SplineMode,
   iwdStrategy: IgnoredWriteDetectionStrategy,
-  userExtraMetadataProvider: UserExtraMetadataProvider,
-  relationHandler: ReadRelationHandler
+  userExtraMetadataProvider: UserExtraMetadataProvider
 ) extends Logging {
 
   private val componentCreatorFactory: ComponentCreatorFactory = new ComponentCreatorFactory
@@ -59,7 +58,7 @@ class LineageHarvester(
   private val opNodeBuilderFactory = new OperationNodeBuilderFactory(userExtraMetadataProvider, componentCreatorFactory, ctx)
   private val pluginRegistry = new PluginRegistryImpl(pathQualifier, ctx.session)
   private val writeCommandExtractor = new PluggableWriteCommandExtractor(pluginRegistry)
-  private val readCommandExtractor = new PluggableReadCommandExtractor(pluginRegistry, relationHandler)
+  private val readCommandExtractor = new PluggableReadCommandExtractor(pluginRegistry)
 
   def harvest(result: Try[Duration]): HarvestResult = {
     log.debug(s"Harvesting lineage from ${ctx.logicalPlan.getClass}")
