+++ core/src/main/scala/za/co/absa/spline/harvester/LineageHarvesterFactory.scala
@@ -19,7 +19,6 @@ package za.co.absa.spline.harvester
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.execution.SparkPlan
-import za.co.absa.spline.harvester.builder.read.ReadRelationHandler
 import za.co.absa.spline.harvester.conf.SplineConfigurer.SplineMode.SplineMode
 import za.co.absa.spline.harvester.extra.UserExtraMetadataProvider
 import za.co.absa.spline.harvester.iwd.IgnoredWriteDetectionStrategy
@@ -29,8 +28,7 @@ import scala.language.postfixOps
 class LineageHarvesterFactory(
   splineMode: SplineMode,
   iwdStrategy: IgnoredWriteDetectionStrategy,
-  userExtraMetadataProvider: UserExtraMetadataProvider,
-  relationHandler: ReadRelationHandler) {
+  userExtraMetadataProvider: UserExtraMetadataProvider) {
 
   def harvester(logicalPlan: LogicalPlan, executedPlan: Option[SparkPlan], session: SparkSession): LineageHarvester =
     new LineageHarvester(
@@ -38,7 +36,6 @@ class LineageHarvesterFactory(
       session.sparkContext.hadoopConfiguration,
       splineMode,
       iwdStrategy,
-      userExtraMetadataProvider,
-      relationHandler
+      userExtraMetadataProvider
     )
 }
