+++ core/src/main/scala/za/co/absa/spline/harvester/conf/DefaultSplineConfigurer.scala
@@ -22,7 +22,6 @@ import org.apache.commons.configuration.{CompositeConfiguration, Configuration, 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SparkSession
 import za.co.absa.commons.config.ConfigurationImplicits
-import za.co.absa.spline.harvester.builder.read.ReadRelationHandler
 import za.co.absa.spline.harvester.conf.SplineConfigurer.SplineMode
 import za.co.absa.spline.harvester.dispatcher.LineageDispatcher
 import za.co.absa.spline.harvester.extra.UserExtraMetadataProvider
@@ -47,11 +46,6 @@ object DefaultSplineConfigurer {
      * Lineage dispatcher used to report lineages
      */
     val LineageDispatcherClass = "spline.lineage_dispatcher.className"
-
-    /**
-     * Relation handler used to deal with proprietary relations
-     */
-    val RelationHandlerClass = "spline.read_relation_handler.className"
 
     /**
      * Strategy used to detect ignored writes
@@ -95,11 +89,6 @@ class DefaultSplineConfigurer(userConfiguration: Configuration) extends SplineCo
   override def queryExecutionEventHandler: QueryExecutionEventHandler =
     new QueryExecutionEventHandler(harvesterFactory, lineageDispatcher)
 
-  protected def relationHandler: ReadRelationHandler = instantiate[ReadRelationHandler](
-    configuration.getString(
-      RelationHandlerClass,
-      "za.co.absa.spline.harvester.builder.read.NoOpReadRelationHandler"))
-
   protected def lineageDispatcher: LineageDispatcher = instantiate[LineageDispatcher](
     configuration.getRequiredString(LineageDispatcherClass))
 
@@ -112,8 +101,8 @@ class DefaultSplineConfigurer(userConfiguration: Configuration) extends SplineCo
   private def harvesterFactory = new LineageHarvesterFactory(
     splineMode,
     ignoredWriteDetectionStrategy,
-    userExtraMetadataProvider,
-    relationHandler)
+    userExtraMetadataProvider
+  )
 
   private def instantiate[T: ClassTag](className: String): T = {
     val interfaceName = scala.reflect.classTag[T].runtimeClass.getSimpleName
