+++ core/src/main/scala/za/co/absa/spline/harvester/plugin/PluginRegistry.scala
@@ -16,54 +16,6 @@
 
 package za.co.absa.spline.harvester.plugin
 
-import io.github.classgraph.ClassGraph
-import org.apache.spark.internal.Logging
-import org.apache.spark.sql.SparkSession
-import za.co.absa.commons.lang.ARM
-import za.co.absa.spline.harvester.qualifier.PathQualifier
-
-import scala.collection.JavaConverters._
-import scala.util.Try
-import scala.util.control.NonFatal
-
 trait PluginRegistry {
   def plugins: Seq[Plugin]
-}
-
-class PluginRegistryImpl(pathQualifier: PathQualifier, session: SparkSession)
-  extends PluginRegistry
-    with Logging {
-
-  private val pluginArgsByType: Map[Class[_], AnyRef] = Map(
-    classOf[PluginRegistry] -> this,
-    classOf[SparkSession] -> session,
-    classOf[PathQualifier] -> pathQualifier
-  )
-
-  override val plugins: Seq[Plugin] = {
-    log.info("Scanning for plugins")
-    val classGraph = new ClassGraph()
-      .enableClassInfo
-      .enableAnnotationInfo
-    for {
-      scanResult <- ARM.managed(classGraph.scan)
-      pluginClass <- scanResult.getClassesImplementing(classOf[Plugin].getName).loadClasses.asScala
-    } yield {
-      log.info(s"Loading plugin: $pluginClass")
-      instantiatePlugin(pluginClass)
-        .recover({
-          case NonFatal(e) => throw new RuntimeException(s"Plugin instantiation failure: $pluginClass", e)
-        })
-        .get
-    }
-  }
-
-  private def instantiatePlugin(pluginClass: Class[_]): Try[Plugin] = Try {
-    val constructors = pluginClass.getConstructors
-    if (constructors.length != 1)
-      sys.error(s"Cannot instantiate a plugin with multiple constructors: $pluginClass")
-    val constr = constructors.head
-    val args = constr.getParameterTypes.map(pt => pluginArgsByType.getOrElse(pt, sys.error(s"Cannot find value for $pt")))
-    constr.newInstance(args: _*).asInstanceOf[Plugin]
-  }
 }
