+++ examples/src/main/resources/spline.properties
@@ -27,4 +27,3 @@
 #
 # spline.mode=DISABLED|REQUIRED|BEST_EFFORT
 # spline.producer.url=http://localhost:8080/spline/producer
-# spline.read_relation_handler.className=za.co.absa.spline.harvester.builder.read.NoOpReadRelationHandler
